requirements.txt
bimpy==0.1.1
dareblopy==0.0.5
dlib==20.0.0
dlutils==0.0.12
geotorch==0.3.0
imageio==2.37.0
matplotlib==3.10.7
more_itertools==10.8.0
mosek==11.0.29
numpy==2.3.3
packaging==25.0
Pillow==11.3.0
pot==0.9.6.post1
pytorch_fid==0.3.0
scikit_learn==1.7.2
scipy==1.16.2
skimage==0.0
tensorflow==2.20.0
torch==2.8.0+cu126
torchsde==0.2.6
torchvision==0.23.0+cu126
tqdm==4.67.1
yacs==0.1.8
